Replace TokenService with IdentityServer
This commit will remove our own implementation of a token service, and
replace it with a service built on IdentityServer.  Instead of using
the normal test setup, we have provided implementations of the
resource and client stores, in order to give a better idea of what it
means to implement an IdP for real production use.

The configuration only supports grant type "client credentials", but
you can easily add new clients.  There is a UI for the interactive
flows, like OIDC, with a single test user "alice" configured.