Update README.md
This commit will update the README.md file to better describe the
scope and purpose of what we are trying to achieve here: a course for
developing secure REST APIs in ASP.NET Core.

The instructions for starting the services are now correct, which
means that the identity service now support command line arguments for
selecting a port.