Include token service in (renamed) solution file
This commit will rename the solution file from "Demo" to "Security" that
better reflects what the source code is all about.  It will also add the
token service project to the solution so that it gets built.

This commit also adds missing JWKS data contract classes to the token
service so that is is complete and builds.

This commit also makes some configuration changes to the token service
so that it is just a little easier to use.