Restructure solution into folders
This commit moves to folders for the different labs, instead of
working with branches.  It turns out that branches is too much work to
maintain.

Lab 6 now also contains a very basic PowerShell module to show how
device flow works, but also as a template for how you can work with a
CLI client for management tasks.  This is a very powerful concept that
means you might not have to develop expensive UI components for some

This commit also moves from .NET Core 1 to version 3.