Add sample code for adding custom claims
This commit will add sample code that describes how you can add custom
USER and CLIENT claims to the access token.  It is commented out, and
is added for reference only.  This is a common scenario that can be a
little tricky to figure out (especially for the client scenario.)